rename rl_chain_base to base and update paths and imports
