+++ libs/langchain/langchain/chains/rl_chain/pick_best_chain.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-import langchain.chains.rl_chain.rl_chain_base as base
+import langchain.chains.rl_chain.base as base
 
 from langchain.callbacks.manager import CallbackManagerForChainRun
 from langchain.chains.base import Chain
