+++ libs/langchain/tests/unit_tests/chains/rl_chain/test_rl_chain_base_embedder.py
@@ -1,4 +1,4 @@
-import langchain.chains.rl_chain as base
+import langchain.chains.rl_chain.base as base
 from test_utils import MockEncoder
 
 import pytest
